ML_Models/Handwritten_Digit_Recognition/requirements.txt
tensorflow>=2.15
matplotlib>=3.8
pillow>=10.0